replace criterion with native cargo bench
